renamed and relocated plugins for better orientation
