+++ project-base/app/assets/js/admin/grapesjs/plugins/mail/image-with-variable.js
