+++ project-base/app/assets/js/admin/grapesjs/plugins/mail/text.js
@@ -1,0 +1,11 @@
+import Translator from 'bazinga-translator';
+import grapesjs from 'grapesjs';
+
+export default grapesjs.plugins.add('mail-text', editor => {
+    editor.Blocks.add('text-ckeditor', {
+        label: Translator.trans('Text'),
+        category: Translator.trans('Basic objects'),
+        attributes: { class: 'gjs-fonts gjs-f-text' },
+        content: { type: 'text-ckeditor', content: Translator.trans('Insert your text here'), activeOnRender: 1 },
+    });
+});
