+++ project-base/app/assets/js/admin/grapesjs/plugins/shared/ckeditor.js
