+++ project-base/app/assets/js/admin/grapesjs/plugins/web/table.js
