+++ project-base/app/assets/js/admin/grapesjs/plugins/web/text-with-image.js
