+++ project-base/app/assets/js/admin/grapesjs/plugins/web/text.js
